Move thread aware advert view away from NoPlayerAdsLoader.
