+++ core/src/main/java/com/novoda/noplayer/internal/exoplayer/NoPlayerAdsLoader.java
@@ -19,7 +19,6 @@ import com.novoda.noplayer.NoPlayer;
 import com.novoda.noplayer.internal.utils.Optional;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
@@ -276,34 +275,4 @@ public class NoPlayerAdsLoader implements AdsLoader, Player.EventListener, Adver
         }
     }
 
-    private static final class MainThreadAwareAdvertView implements AdvertView {
-
-        private final AdvertView advertView;
-        private final Handler handler;
-
-        private MainThreadAwareAdvertView(AdvertView advertView, Handler handler) {
-            this.advertView = advertView;
-            this.handler = handler;
-        }
-
-        @Override
-        public void setup(final List<AdvertBreak> advertBreaks, final AdvertInteractionListener advertInteractionListener) {
-            handler.post(new Runnable() {
-                @Override
-                public void run() {
-                    advertView.setup(new ArrayList<>(advertBreaks), advertInteractionListener);
-                }
-            });
-        }
-
-        @Override
-        public void removeMarker(final AdvertBreak advertBreak) {
-            handler.post(new Runnable() {
-                @Override
-                public void run() {
-                    advertView.removeMarker(advertBreak); // TODO: Maybe we should create a copy?
-                }
-            });
-        }
-    }
 }
